requirements.txt
python-dotenv
sendgrid==5.6.0
pandas
